requirements.txt
aiogram
aiohttp
tortoise-orm[asyncpg]
rich
uvloop
ujson
python-rapidjson
aerich==0.6.3
validators